+++ src/Exceptions/MaxRetriesReachedException.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Martingalian\Core\Exceptions;
-
-use Exception;
-
-final class MaxRetriesReachedException extends Exception {}